Fix logbook username not appearing consistently (#6230)
